renaming + refactoring
