+++ src/components/App.js
@@ -1,4 +1,4 @@
-import Router from "./Router";
+import Router from "../Router";
 import "./App.css";
 
 function App() {
