+++ src/index.js
@@ -1,9 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
 import { BrowserRouter } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "react-query";
-import { ErrorBoundary } from "./components";
+import { App, ErrorBoundary } from "./components";
 
 const queryClient = new QueryClient({
   defaultOptions: {
