+++ src/pages/Team.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { Pagination, Breadcrumbs, DateFilter } from "../components";
-import { ErrorModule } from "../components";
+import { ErrorComponent } from "../components";
 import { useGetTeamMatches } from "../logic/query";
-import { parseMatchData } from "../logic/pageUtils";
+import { parseMatchData, paginateItems } from "../logic/pageUtils";
 import { filterMatchesByDate, findTeam } from "../logic/pageUtils";
 
 const Team = () => {
@@ -23,18 +23,12 @@ const Team = () => {
 
   const filteredMatches = filterMatchesByDate(data, dateRange);
 
-  const totalItems = filteredMatches.length;
-  const itemsPerPage = 20;
-
-  const pageMatches = filteredMatches.slice(
-    activePage * itemsPerPage - itemsPerPage,
-    activePage * itemsPerPage
-  );
+  const pageMatches = paginateItems(filteredMatches, activePage, 20);
 
   return (
     <>
-      {error !== null ? (
-        <ErrorModule error={error} />
+      {error ? (
+        <ErrorComponent error={error} />
       ) : (
         <>
           <Breadcrumbs openedElement={teamName} />
@@ -64,8 +58,8 @@ const Team = () => {
           <Pagination
             activePage={activePage}
             setActivePage={setActivePage}
-            totalItems={totalItems}
-            itemsPerPage={itemsPerPage}
+            totalItems={filteredMatches.length}
+            itemsPerPage={20}
           />
         </>
       )}
